update indexed row suite
